+++ tfinal/src/Point.cpp
@@ -7,7 +7,7 @@ Point::Point() {
     this->z = 0;
 }
 
-Point::Point(float x, float y, float z) {
+Point::Point(double x, double y, double z) {
     this->x = x;
     this->y = y;
     this->z = z;
@@ -17,61 +17,61 @@ Point Point::copy() {
     return Point(x, y, z);
 }
 
-float Point::getX() {
+double Point::getX() {
     return x;
 }
 
-float Point::getY() {
+double Point::getY() {
     return y;
 }
 
-float Point::getZ() {
+double Point::getZ() {
     return z;
 }
 
-void Point::setX(float x) {
+void Point::setX(double x) {
     this->x = x;
 }
 
-void Point::setY(float y) {
+void Point::setY(double y) {
     this->y = y;
 }
 
-void Point::setZ(float z) {
+void Point::setZ(double z) {
     this->z = z;
 }
 
-void Point::translate(float x, float y, float z) {
+void Point::translate(double x, double y, double z) {
     this->x += x;
     this->y += y;
     this->z += z;
 }
 
-void Point::rotateX(float ang) {
-    float yTemp = cos(ang) * this->y - sin(ang) * this->z;
-    float zTemp = sin(ang) * this->y + cos(ang) * this->z;
+void Point::rotateX(double ang) {
+    double yTemp = cos(ang) * this->y - sin(ang) * this->z;
+    double zTemp = sin(ang) * this->y + cos(ang) * this->z;
 
     this->y = yTemp;
     this->z = zTemp;
 }
 
-void Point::rotateY(float ang) {
-    float xTemp = cos(ang) * this->x + sin(ang) * this->z;
-    float yTemp = cos(ang) * this->z - sin(ang) * this->x;
+void Point::rotateY(double ang) {
+    double xTemp = cos(ang) * this->x + sin(ang) * this->z;
+    double yTemp = cos(ang) * this->z - sin(ang) * this->x;
 
     this->x = xTemp;
     this->z = yTemp;
 }
 
-void Point::rotateZ(float ang) {
-    float xTemp = cos(ang) * this->x - sin(ang) * this->y;
-    float yTemp = sin(ang) * this->x + cos(ang) * this->y;
+void Point::rotateZ(double ang) {
+    double xTemp = cos(ang) * this->x - sin(ang) * this->y;
+    double yTemp = sin(ang) * this->x + cos(ang) * this->y;
 
     this->x = xTemp;
     this->y = yTemp;
 }
 
-void Point::project(float d) {
+void Point::project(double d) {
     this->x = this->x * d / this->z;
     this->y = this->y * d / this->z;
     this->z = 0;
